Automatically generate system call prototypes for the Linux kernel.
Inside of the Linux kernel we have the possibility of simply using the
same prototypes as what we use for userspace, as long as our system call
table has wrappers to invoke them in the right way. Already check in the
generator for the header files for our prototypes. Some differences
compared to userspace:

- We don't want them to be static inline. Move this into the system call
  implementation generator.
- Don't use _Noreturn. The implementations of these functions need to
  invoke other functions that are not marked that way.
- Add __user annotations to pointers going to userspace.

I'm now working on patching up all of the system calls to use these
prototypes. When this is completed, I'll work on automatically
generating a system call table.