Add AT_SYSINFO_EHDR.
This is going to be the auxiliary vector entry that we're going to use
to point to the vDSO. In our case we're going to make pretty advanced
use of the vDSO, because we want to have it invoked for every individual
system call. This makes emulation a lot easier.