Remove mem_lock() and mem_unlock().
As discussed on IRC, having them doesn't make a lot of sense in our
environment. According to this article:

https://eklitzke.org/mlock-and-mlockall

They are mainly useful for applications that:

- Are real-time, or
- Store secure/privileged information, like cryptographic keys.

CloudABI is not focussed on real-time performance. I think the second
use case is also not that strong. A better way of solving this is to use
an encrypted swap partition.