Remove socket creation, listening, binding and connecting.
The problem with these interfaces is that they don't allow existing
socket file descriptors to be injected. For example, it is impossible to
push an IPv6 connection into an application that has created its own
UNIX socket and is accepting connections.

Now that Flower is becoming more complete, there is no real need anymore
to keep this functionality around. Flower allows programs to wait for
arbitrary connections.

At some point we should also remove sock_accept(), but there are still
some things out there (including demo apps) that make use of it.