Remove the poll_fd() system call and everything associated with it.
In an attempt to bring event driven programming to a higher level, let's
standardize on an event loop in the C library. The goal will be to add
an implementation of libuv to cloudlibc. Initially, it will be built on
top of plain poll().

While inefficient, this will only be the first step. Once we've got it
working, we should design a kernel-space polling framework that exactly
fits libuv's requirements. This will likely mean process descriptors may
also disappear at some point, as subprocesses will be managed by the
polling framework, as opposed to being a file descriptor type.