Add a UUID field to the auxiliary vector.
cloudlibc already has the program_getuuid() function for some time,
which returns a UUID of the running process. Let's go ahead and move the
generation of this number into the runtime.

This will allow us to have administrative tools that identify processes
by UUID, which is going to be quite useful.