Rewrite the ARMv6-on-64bit vDSO to be less demanding on registers.
Instead of hoping that all return value addresses are retained in
registers, just push them on the stack. Put the system call number in
r0 and the padded buffer pointer in r2.

This makes it possible for me to run ARMv6 executables on FreeBSD/arm64
without as many hacks as before. Forking now works as well.